refactor code -> put main files in tlidb folder
