+++ tlidb/examples/algorithms/EncoderAlgorithm.py
@@ -1,5 +1,5 @@
-from examples.utils import move_to
-from examples.losses import initialize_loss
+from tlidb.examples.utils import move_to
+from tlidb.examples.losses import initialize_loss
 from .algorithm import Algorithm
 import torch
