+++ tlidb/examples/algorithms/EncoderDecoderAlgorithm.py
@@ -1,7 +1,7 @@
-from examples.utils import move_to
+from tlidb.examples.utils import move_to
 from .algorithm import Algorithm
 
-class DecoderAlgorithm(Algorithm):
+class EncoderDecoderAlgorithm(Algorithm):
     def __init__(self, config, datasets):
         super().__init__(config, datasets)
         self.generation_config = config.generation_config
@@ -26,23 +26,26 @@ class DecoderAlgorithm(Algorithm):
 
         # prepare inputs for generation if necessary
         if self.requires_metric_calculation():
-            X_generate = self.model.transform_generation_inputs(X)
+            self.model.tokenizer.padding_side="left"
+            X_generate = self.model.transform_inputs(X).input_ids
             X_generate = move_to(X_generate, self.device)
+            self.model.tokenizer.padding_side="right"
 
-        X, lm_labels = self.model.transform_LM_inputs(X,y_true)
+        X = self.model.transform_inputs(X)
+        lm_labels = self.model.transform_outputs(y_true)
+
         X['lm_labels'] = lm_labels
         X = move_to(X, self.device)
 
-        # track number of tokens in the batch
-        num_batch_tokens = X['attention_mask'].sum().item()
+        num_batch_tokens = (X['lm_labels'] != -100).sum().item() 
 
         loss = self.model(**X)
 
         # generate predictions and convert to labels if necessary
         if self.requires_metric_calculation():
             # generate predictions
-            outputs = self.model.generate(X_generate, metadata['task_metadata']['max_decode_tokens'], **self.generation_config)
-
+            outputs = self.model.generate(X_generate, **self.generation_config)
+            
             # task-specific postprocessing
             y_pred, y_true = getattr(self, f"_{metadata['task_metadata']['type']}_postprocessing")(outputs, y_true, metadata)
 
@@ -54,7 +57,7 @@ class DecoderAlgorithm(Algorithm):
             'y_pred': y_pred,
             'y_true': y_true,
             'metadata': metadata,
-            'batch_loss_divisor': num_batch_tokens, # used for averaging loss
+            'batch_loss_divisor': num_batch_tokens, # number of tokens used for language modeling
             "objective": {
                 "loss_name": "LM_cross_entropy",
                 "loss_value": loss.item()*num_batch_tokens}
@@ -73,26 +76,30 @@ class DecoderAlgorithm(Algorithm):
         return X, y_true, metadata
 
     def _classification_postprocessing(self, outputs, y_true, metadata):
+        outputs = self.model.batch_decode(outputs)
         y_true = self.convert_strings_to_labels(metadata['labels'], y_true)
         assert(all(y_true != -1)),str(y_true)
         y_pred = self.convert_strings_to_labels(metadata['labels'], outputs)
+        
         return y_pred, y_true
 
     def _multioutput_classification_preprocessing(self, X, y_true, metadata):
         return X, y_true, metadata
 
     def _multioutput_classification_postprocessing(self, outputs, y_true, metadata):
+        outputs = self.model.batch_decode(outputs)
         y_true = self.convert_strings_to_labels(metadata['labels'], y_true)
         assert(all(y_true != -1)),str(y_true)
         y_pred = self.convert_strings_to_labels(metadata['labels'], outputs)
         return y_pred, y_true
-    
+
     def _multilabel_classification_preprocessing(self, X, y_true, metadata):
         # format y_true into a string of labels
         y_true = [" and ".join([metadata['task_metadata']['class_to_natural_language_map'][c] for c in sample]) for sample in y_true]
         return X, y_true, metadata
 
     def _multilabel_classification_postprocessing(self, outputs, y_true, metadata):
+        outputs = self.model.batch_decode(outputs)
         # convert model outputs to mutlilabel format
         y_pred = []
         for output in outputs:
@@ -126,13 +133,14 @@ class DecoderAlgorithm(Algorithm):
         return X, y_true, metadata
 
     def _span_extraction_postprocessing(self, outputs, y_true, metadata):
-        y_pred = outputs
+        y_pred = self.model.batch_decode(outputs)
         return y_pred, y_true
 
     def _multiple_choice_preprocessing(self, X, y_true, metadata):
         return X, y_true, metadata
 
     def _multiple_choice_postprocessing(self, outputs, y_true, metadata):
+        outputs = self.model.batch_decode(outputs)
         num_choices = metadata['task_metadata']['num_choices']
         metadata['labels'] = [str(i) for i in range(num_choices)]
         y_true = self.convert_strings_to_labels(metadata['labels'], y_true)
@@ -142,7 +150,7 @@ class DecoderAlgorithm(Algorithm):
 
     def _response_generation_preprocessing(self, X, y_true, metadata):
         return X, y_true, metadata
-    
+
     def _response_generation_postprocessing(self, outputs, y_true, metadata):
-        y_pred = outputs
+        y_pred = self.model.batch_decode(outputs)
         return y_pred, y_true
