+++ tlidb/examples/algorithms/initializer.py
@@ -1,6 +1,6 @@
-from examples.algorithms.EncoderAlgorithm import EncoderAlgorithm
-from examples.algorithms.EncoderDecoderAlgorithm import EncoderDecoderAlgorithm
-from examples.algorithms.DecoderAlgorithm import DecoderAlgorithm
+from tlidb.examples.algorithms.EncoderAlgorithm import EncoderAlgorithm
+from tlidb.examples.algorithms.EncoderDecoderAlgorithm import EncoderDecoderAlgorithm
+from tlidb.examples.algorithms.DecoderAlgorithm import DecoderAlgorithm
 
 def initialize_algorithm(config, datasets):
     """Load an algorithm of type Algorithm
